feat(v1Router):[+] Moved all v1 endpoints to v1Router file
